refactor(canvas): split target action modules
- extract canvas target action creation, layout, grouping, and update helpers

- keep the public action catalog wired through the original entry point
